Move generated typescript location
