+++ generators/typescript-fetch/tests/multi-path-variables.test.ts
@@ -4,7 +4,7 @@ import fetch from 'node-fetch';
 import { describe, beforeAll, afterEach, afterAll, it, expect } from 'vitest';
 import { toFetchApi, toFetchOperation } from '../src';
 import type { FetchImplementation } from '../src';
-import operations from './multi-path-variables/operations';
+import operations from './generated/multi-path-variables/operations';
 
 const baseDomain = 'http://localhost/';
 const fetchImpl: FetchImplementation<unknown> = (url, params) =>
