+++ generators/typescript-msw/npm/npm-typescript-msw.esproj
@@ -22,7 +22,7 @@
     </ItemGroup>
 
     <Target Name="GenerateFromSchema" BeforeTargets="Generation">
-        <Exec Command="pnpm openapi-codegen-typescript -o ./openapi-codegen-options.yaml ../../../schemas/%(TestedSchema.Identity).yaml tests/%(TestedSchema.Identity)/" />
+        <Exec Command="pnpm openapi-codegen-typescript -o ./openapi-codegen-options.yaml ../../../schemas/%(TestedSchema.Identity).yaml tests/generated/%(TestedSchema.Identity)/" />
     </Target>
 
 </Project>
