+++ generators/typescript-msw/npm/tests/petstore.test.ts
@@ -2,8 +2,8 @@ import { setupServer } from 'msw/node';
 import fetch from 'node-fetch';
 import { describe, beforeAll, afterEach, afterAll, it, expect } from 'vitest';
 import { toMswHandler, toMswResponse } from '../src';
-import type { NewPet } from './petstore/models';
-import operations from './petstore/operations';
+import type { NewPet } from './generated/petstore/models';
+import operations from './generated/petstore/operations';
 
 const baseDomain = 'http://localhost/';
 const findPets = toMswHandler(operations.findPets, { baseDomain });
